Add initial config that will be edited later in tutorial
This initial config is based on our boilerplate Node config
(and for GH/BB users there is the option to do the express
config setup and select this template). It does not yet have
the circleci tests command added.

We are "reverting" to this old/incomplete config so GitLab
users (who do not have the express config setup option) can
still follow along the tutorial and edit the config as they
go.